Fix X11 Pixel truncation issues
- Change all Default*Pixel variables from int to unsigned long to match
  X11 Pixel typedef (unsigned long on 64-bit systems)
- Update initDefaultColor() to accept unsigned long parameters
- Update GLSetBGColorPixel() to accept unsigned long pixel parameter
- Change prev*Pixel cache variables to unsigned long with ~0UL sentinel
- Fix extern declarations in cl_field.c and cl_menu.c

This prevents potential color corruption when X server returns pixel IDs
larger than INT_MAX (2^31-1), though rare in practice with typical 24-bit
color depths.

Fixes 21 -Wshorten-64-to-32 warnings (394 -> 373).

+++ src/viola/cl_field.c
@@ -1200,10 +1200,10 @@ long meth_field_initialize(VObj* self, Packet* result, int argc, Packet argv[]) 
      */
     cip = GET__classInfo(self);
     if (cip == &class_slider) {
-        extern int DefaultBGPixel_darkest;
+        extern unsigned long DefaultBGPixel_darkest;
         GLSetBGColorPixel(self, DefaultBGPixel_darkest);
     } else if (cip == &class_txtEdit) {
-        extern int DefaultBGPixel_darker;
+        extern unsigned long DefaultBGPixel_darker;
         GLSetBGColorPixel(self, DefaultBGPixel_darker);
     } else {
         cp = GET_BGColor(self);
